Added strafe correction opMode.
Allows for a perfect strafe in a user defined direction. Finally you can take out your autonomous frustrations on your robot without adversely affecting its course! Kick the robot while going sideways (mean tho) and watch as it perfectly adjusts to its original direction. Or, if you are nice to robots, press init an angle, twist the robot by up to 90 degrees, and press play and watch as the robot returns to the the init angle while strafing within seconds!